+++ src/Model/Feed/FeedItem.php
@@ -4,22 +4,26 @@ namespace App\Model\Feed;
 
 final class FeedItem
 {
-    public readonly int $id;
-    public string $description = '';
-    public string $excerpt = '';
-    public string $image = '';
-    public string $ticketUrl = '';
-    public string $title = '';
-    public string $url = '';
-    public ?\DateTimeImmutable $end = null;
-    public ?\DateTimeImmutable $start = null;
-    /** @var array<FeedItemOccurrence> */
-    public array $occurrences = [];
-    public ?FeedItemLocation $location = null;
-    public string $feedId = '';
-    public string $price = '';
-    /** @var array<string> */
-    public array $tags = [];
+    public function __construct(
+        public readonly int $id,
+        public string $description = '',
+        public string $excerpt = '',
+        public string $image = '',
+        public string $ticketUrl = '',
+        public string $title = '',
+        public string $url = '',
+        public ?\DateTimeImmutable $end = null,
+        public ?\DateTimeImmutable $start = null,
+        /** @var array<FeedItemOccurrence> */
+        public array $occurrences = [],
+        public ?FeedItemLocation $location = null,
+        public string $feedId = '',
+        public string $price = '',
+        /** @var array<string> */
+        public array $tags = [],
+    )
+    {
+    }
 
     public function __toString(): string
     {