+++ CHANGELOG.md
@@ -21,3 +21,4 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 - Fixtures
 - Typed feed configuration objects
 - Feeds default mapper service
+- Added phpunit tests